test(core): split hmr perf tests
